pyproject.toml
[tool.poetry]
name = "ai-agents"
version = "0.1.0"
description = ""
authors = ["Andres Velez <andres.velez.e@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
tavily-python = "^0.5.0"
langchain = "^0.3.18"
langchain-community = "^0.3.17"
python-dotenv = "^1.0.1"
openai = "^1.61.1"
serpapi = "^0.1.5"
google-search-results = "^2.4.2"
wikipedia = "^1.4.0"
pytz = "^2025.1"
panel = "^1.6.0"
langchain-openai = "^0.3.4"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
